Extract assertions specific to Option<T, Exception> to separate class
