+++ src/FluentAssertions.Optional/OptionEitherAssertions.cs
@@ -1,31 +1,8 @@
-﻿using System;
-using System.Collections.Generic;
-using FluentAssertions.Execution;
-using FluentAssertions.Specialized;
+﻿using FluentAssertions.Execution;
 using Optional;
 
 namespace FluentAssertions.Optional
 {
-    // public static class OptionEitherAssertionExtensions
-    // {
-    //     public static ExceptionAssertions<TException> HaveException<T, TException>(
-    //         this OptionEitherAssertions<T, TException> assertions,
-    //         string because = "",
-    //         params object[] becauseArgs)
-    //         where TException : Exception
-    //     {
-    //         Execute.Assertion
-    //             .ForCondition(!assertions._subject.HasValue)
-    //             .BecauseOf(because, becauseArgs)
-    //             .FailWith("Expected {context:option} to be None{reason} but found {0}.", assertions._subject);
-    //         
-    //         TException exception = null;
-    //         assertions._subject.MatchNone(ex => exception = ex);
-    //         
-    //         return new ExceptionAssertions<TException>(new []{exception});
-    //     }
-    // }
-    
     public class OptionEitherAssertions<T, TException>
     {
         internal readonly Option<T, TException> _subject;
@@ -61,29 +38,6 @@ namespace FluentAssertions.Optional
                 .FailWith("Expected {context:Option} not to be {0}{reason}, but found {1}.", other, _subject);
 
             return new AndConstraint<OptionEitherAssertions<T, TException>>(this);
-        }
-        
-        [CustomAssertion]
-        public ExceptionAssertions<TExpectedException> HaveException<TExpectedException>(string because = "", params object[] becauseArgs) where TExpectedException : Exception, TException
-        {
-            Execute.Assertion
-                .ForCondition(!_subject.HasValue)
-                .BecauseOf(because, becauseArgs)
-                .FailWith("Expected {context:option} to have exception{reason} but found {0}.", _subject);
-
-            TException _exception = default(TException);
-            _subject.MapException(ex => _exception = ex);
-            
-            return new ExceptionAssertions<TExpectedException>(new List<TExpectedException>{(TExpectedException) _exception});
-        }
-        
-        [CustomAssertion]
-        public void NotHaveException(string because = "", params object[] becauseArgs)
-        {
-            Execute.Assertion
-                .ForCondition(_subject.HasValue)
-                .BecauseOf(because, becauseArgs)
-                .FailWith("Expected {context:option} not to have exception{reason} but found {0}.", _subject);
         }
         
         [CustomAssertion]
